feat: remove changeset and move code to diff file.
