--- .changeset/humble-meteors-poke.md
@@ -1,5 +1,0 @@
----
-
----
-
-feat: add interface to get selected banks
